Добавил валидации для категорий и продуктов

+++ app/Http/Controllers/Admin/CategoryController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
+use App\Http\Requests\CategoryRequest;
 use App\Models\Category;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
@@ -36,11 +37,14 @@ class CategoryController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)  //Процесс добавления новой формы
+    public function store(CategoryRequest $request)  //Процесс добавления новой формы
     {
-        $path = $request->file('image')->store('categories');
         $params = $request->all();
-        $params['image'] = $path;
+        unset($params['image']);
+        if ($request->has('image')){
+            $path = $request->file('image')->store('categories');
+            $params['image'] = $path;
+        }
         Category::create($params);
         return redirect() ->route('admin.categories.index');
     }
@@ -75,12 +79,15 @@ class CategoryController extends Controller
      * @param  \App\Models\Category  $category
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, Category $category) //Процесс редактирования данных
+    public function update(CategoryRequest $request, Category $category) //Процесс редактирования данных
     {
-        Storage::delete($category->image);
-        $path = $request->file('image')->store('categories');
         $params = $request->all();
-        $params['image'] = $path;
+        unset($params['image']);
+        if ($request->has('image')) {
+            Storage::delete($category->image);
+            $path = $request->file('image')->store('categories');
+            $params['image'] = $path;
+        }
         $category->update($params);
         return redirect() ->route('admin.categories.index');
     }
