+++ app/Http/Controllers/Admin/ProductController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
+use App\Http\Requests\ProductRequest;
 use App\Models\Category;
 use App\Models\Product;
 use Illuminate\Http\Request;
@@ -39,11 +40,14 @@ class ProductController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request) //Процесс добавления новой формы
+    public function store(ProductRequest $request) //Процесс добавления новой формы
     {
-        $path = $request->file('image')->store('products');
         $params = $request->all();
-        $params['image'] = $path;
+        unset($params['image']);
+        if ($request->has('image')){
+            $path = $request->file('image')->store('products');
+            $params['image'] = $path;
+        }
         Product::create($params);
         return redirect() ->route('admin.products.index');
     }
@@ -78,12 +82,15 @@ class ProductController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, Product $product) //Процесс редактирования данных
+    public function update(ProductRequest $request, Product $product) //Процесс редактирования данных
     {
-        Storage::delete($product->image);
-        $path = $request->file('image')->store('products');
         $params = $request->all();
-        $params['image'] = $path;
+        unset($params['image']);
+        if ($request->has('image')) {
+            Storage::delete($product->image);
+            $path = $request->file('image')->store('products');
+            $params['image'] = $path;
+        }
         $product->update($params);
         return redirect() ->route('admin.products.index');
     }
